test: refactor test hierarchy
